Separate out legacy/v3 identity in xmtp_mls (#681)
Separate old identity and new identity into mod/feature
