+++ xmtp_mls/Cargo.toml
@@ -10,6 +10,7 @@ path = "src/bin/update-schema.rs"
 
 [features]
 default = ["types", "native"]
+xmtp-id = []
 grpc = ["xmtp_proto/grpc"]
 native = ["libsqlite3-sys/bundled-sqlcipher-vendored-openssl"]
 types = []
@@ -55,3 +56,4 @@ tempfile = "3.5.0"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
 xmtp_api_grpc = { path = "../xmtp_api_grpc" }
 xmtp_id = { path = "../xmtp_id", features = ["test-utils"] }
+
